fix: remove cast for iterator

+++ 0x03-python-data_structures/100-print_python_list_info.c
@@ -14,7 +14,7 @@ void print_python_list_info(PyObject *p)
 
 	printf("[*] Size of the Python List = %ld\n", lp->ob_base.ob_size);
 	printf("[*] Allocated = %ld\n", lp->allocated);
-	for (i = 0; i < lp->ob_base.ob_size; i++)
+	for (i = 0; i < (int)lp->ob_base.ob_size; i++)
 	{
 		printf("Element %d: %s\n", i, lp->ob_item[i]->ob_type->tp_name);
 	}
